requirements.txt
beautifulsoup4==4.12.3
Jinja2==3.1.4
kaleido==0.2.1
matplotlib==3.9.0
matplotlib-inline==0.1.7
numpy==1.26.4
pandas==2.2.2
plotly==5.22.0
python-dotenv==1.0.1
redshift-connector==2.1.1